# openapi.yaml
openapi: 3.0.1
info:
  title: Month and Math API
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /month/{monthNumber}:
    get:
      tags:
        - math-and-month-controller
      operationId: monthConverter
      parameters:
        - name: monthNumber
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Month'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'

  /randomMonth:
    get:
      tags:
        - math-and-month-controller
      operationId: randomMonthGetter
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Month'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /add:
    post:
      tags:
        - math-and-month-controller
      operationId: addNums
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /subtract:
    post:
      tags:
        - math-and-month-controller
      operationId: subtractNums
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /multiply:
    post:
      tags:
        - math-and-month-controller
      operationId: multiplyNums
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
  /divide:
    post:
      tags:
        - math-and-month-controller
      operationId: divideNums
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MathSolution'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MathSolution'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomErrorResponse'
                  - type: array
                    items:
                      $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
    MathSolution:
      required:
        - operand1
        - operand2
      type: object
      properties:
        operand1:
          type: integer
          format: int32
        operand2:
          type: integer
          format: int32
        operation:
          type: string
        answer:
          type: integer
          format: int32
    Month:
      required:
        - name
        - number
      type: object
      properties:
        number:
          maximum: 12
          minimum: 1
          type: integer
          format: int32
        name:
          type: string